TodoApp/requirements.txt
python3
django=3
djangorestframework
corsheaders
